Implemented PlaceholderMorph (but haven't started using it yet).
The idea is that two poses might both want to include the same morph,
so if one steals a morph from the other, it can leave a PlaceholderMorph
to mark the spot where it stole the morph from. Then when the user is
looking at that spot later, he can click on the PlaceholderMorph to have
the real morph return.