Show a mirror's __proto__ slot in in the object header, rather than as part of the contents.
I'm not completely sure that this is better, but I think it might be. An object can only have one parent, plus it behaves differently from other slots; treating it specially makes sense. And I find that I'm either looking specifically for an object's parent (in which case it helps to have it in a special place), or else I don't care about it at all and I just want to see the object's other slot (in which case it also helps to have the parent slot in a separate place).

Also, changed the slot's title to "inherits from" rather than "__proto__". Should be clearer.