I'm not sure this is the right thing to do, but I made okToBeGrabbedBy take the owner's openForDragAndDrop into account.
This does allow me to make, e.g., the contents of a TreeNodeMorph ungrabbable (which is nice for keeping things in place - otherwise we'd have to tell each of the individual submorphs to suppressGrabbing, plus that wouldn't even work right in cases where we had picked up the content morph and moved it somewhere else, like the world, where it *should* be grabbable). But I'm not convinced yet that this isn't going have unintended consequences.